Build maps from res://maps/ folder during runtime
Still a big work in progress
Currently, I don't know why, but it seg faults whenever it tries to
build a map

It creates a level config for each .map files which loads the bulid
scene. The build scene then finds the path from a dictionary in the
ModLoader and adds it. Then the map gets built and currently causes
a seg fault.